web: refactor 6/7 node page
