+++ web/public/assets/js/app/node-page/single-node-table.js
@@ -1,0 +1,139 @@
+/*
+ * Copyright © 2025-26 l5yth & contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * Render a one-row table summarising the selected node's metrics.
+ *
+ * @module node-page/single-node-table
+ */
+
+import { escapeHtml } from '../utils.js';
+import { renderNodeLongNameLink } from '../node-rendering.js';
+import {
+  fmtAlt,
+  fmtHumidity,
+  fmtPressure,
+  fmtTemperature,
+  fmtTx,
+} from '../short-info-telemetry.js';
+import {
+  formatBattery,
+  formatCoordinate,
+  formatDurationSeconds,
+  formatHardwareModel,
+  formatRelativeSeconds,
+  formatVoltage,
+} from '../node-page-charts.js';
+import { numberOrNull, stringOrNull } from '../value-helpers.js';
+import { renderRoleAwareBadge } from './neighbor-rendering.js';
+
+/**
+ * Render a condensed node table containing a single entry.
+ *
+ * @param {Object} node Normalised node payload.
+ * @param {Function} renderShortHtml Badge rendering implementation.
+ * @param {number} [referenceSeconds] Optional reference timestamp for relative metrics.
+ * @returns {string} HTML markup for the node table or an empty string.
+ */
+export function renderSingleNodeTable(node, renderShortHtml, referenceSeconds = Date.now() / 1000) {
+  if (!node || typeof node !== 'object' || typeof renderShortHtml !== 'function') {
+    return '';
+  }
+  const nodeId = stringOrNull(node.nodeId ?? node.node_id) ?? '';
+  const shortName = stringOrNull(node.shortName ?? node.short_name) ?? null;
+  const longName = stringOrNull(node.longName ?? node.long_name);
+  const protocol = stringOrNull(node.protocol) ?? null;
+  const longNameLink = renderNodeLongNameLink(longName, nodeId, { protocol });
+  const role = stringOrNull(node.role) ?? 'CLIENT';
+  const numericId = numberOrNull(node.nodeNum ?? node.node_num ?? node.num);
+  const badgeSource = node.rawSources?.node && typeof node.rawSources.node === 'object'
+    ? node.rawSources.node
+    : node;
+  const badgeHtml = renderRoleAwareBadge(renderShortHtml, {
+    shortName,
+    longName,
+    role,
+    identifier: nodeId || null,
+    numericId,
+    source: badgeSource,
+  });
+  const hardware = formatHardwareModel(node.hwModel ?? node.hw_model);
+  const battery = formatBattery(node.battery ?? node.battery_level);
+  const voltage = formatVoltage(node.voltage ?? node.voltageReading);
+  const uptime = formatDurationSeconds(node.uptime ?? node.uptime_seconds ?? node.uptimeSeconds);
+  const channelUtil = node.channel_utilization ?? node.channelUtilization ?? null;
+  const channel = fmtTx(channelUtil, 3);
+  const airUtil = fmtTx(node.airUtil ?? node.air_util_tx ?? node.airUtilTx ?? null, 3);
+  const temperature = fmtTemperature(node.temperature ?? node.temp);
+  const humidity = fmtHumidity(node.humidity ?? node.relative_humidity ?? node.relativeHumidity);
+  const pressure = fmtPressure(node.pressure ?? node.barometric_pressure ?? node.barometricPressure);
+  const latitude = formatCoordinate(node.latitude ?? node.lat);
+  const longitude = formatCoordinate(node.longitude ?? node.lon);
+  const altitude = fmtAlt(node.altitude ?? node.alt, 'm');
+  const lastSeen = formatRelativeSeconds(node.lastHeard ?? node.last_heard, referenceSeconds);
+  const lastPosition = formatRelativeSeconds(node.positionTime ?? node.position_time, referenceSeconds);
+
+  return `
+    <div class="nodes-table-wrapper">
+      <table class="nodes-detail-table" aria-label="Selected node details">
+        <thead>
+          <tr>
+            <th class="nodes-col nodes-col--node-id">Node ID</th>
+            <th class="nodes-col nodes-col--short-name">Short</th>
+            <th class="nodes-col nodes-col--long-name">Long Name</th>
+            <th class="nodes-col nodes-col--last-seen">Last Seen</th>
+            <th class="nodes-col nodes-col--role">Role</th>
+            <th class="nodes-col nodes-col--hw-model">HW Model</th>
+            <th class="nodes-col nodes-col--battery">Battery</th>
+            <th class="nodes-col nodes-col--voltage">Voltage</th>
+            <th class="nodes-col nodes-col--uptime">Uptime</th>
+            <th class="nodes-col nodes-col--channel-util">Channel Util</th>
+            <th class="nodes-col nodes-col--air-util-tx">Air Util Tx</th>
+            <th class="nodes-col nodes-col--temperature">Temperature</th>
+            <th class="nodes-col nodes-col--humidity">Humidity</th>
+            <th class="nodes-col nodes-col--pressure">Pressure</th>
+            <th class="nodes-col nodes-col--latitude">Latitude</th>
+            <th class="nodes-col nodes-col--longitude">Longitude</th>
+            <th class="nodes-col nodes-col--altitude">Altitude</th>
+            <th class="nodes-col nodes-col--last-position">Last Position</th>
+          </tr>
+        </thead>
+        <tbody>
+          <tr>
+            <td class="mono nodes-col nodes-col--node-id">${escapeHtml(nodeId)}</td>
+            <td class="nodes-col nodes-col--short-name">${badgeHtml}</td>
+            <td class="nodes-col nodes-col--long-name">${longNameLink}</td>
+            <td class="nodes-col nodes-col--last-seen">${escapeHtml(lastSeen)}</td>
+            <td class="nodes-col nodes-col--role">${escapeHtml(role)}</td>
+            <td class="nodes-col nodes-col--hw-model">${escapeHtml(hardware)}</td>
+            <td class="nodes-col nodes-col--battery">${escapeHtml(battery ?? '')}</td>
+            <td class="nodes-col nodes-col--voltage">${escapeHtml(voltage ?? '')}</td>
+            <td class="nodes-col nodes-col--uptime">${escapeHtml(uptime)}</td>
+            <td class="nodes-col nodes-col--channel-util">${escapeHtml(channel ?? '')}</td>
+            <td class="nodes-col nodes-col--air-util-tx">${escapeHtml(airUtil ?? '')}</td>
+            <td class="nodes-col nodes-col--temperature">${escapeHtml(temperature ?? '')}</td>
+            <td class="nodes-col nodes-col--humidity">${escapeHtml(humidity ?? '')}</td>
+            <td class="nodes-col nodes-col--pressure">${escapeHtml(pressure ?? '')}</td>
+            <td class="nodes-col nodes-col--latitude">${escapeHtml(latitude)}</td>
+            <td class="nodes-col nodes-col--longitude">${escapeHtml(longitude)}</td>
+            <td class="nodes-col nodes-col--altitude">${escapeHtml(altitude ?? '')}</td>
+            <td class="mono nodes-col nodes-col--last-position">${escapeHtml(lastPosition)}</td>
+          </tr>
+        </tbody>
+      </table>
+    </div>
+  `;
+}
